pyproject.toml
[tool.poetry]
name = "starlette-sqlalchemy"
version = "0.1.0"
description = ""
authors = ["Jonathan Kim <jkimbo@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
uvicorn = {extras = ["standard"], version = "^0.15.0"}
SQLAlchemy = {extras = ["mypy"], version = "^1.4.22"}
alembic = "^1.7.1"
strawberry-graphql = {extras = ["asgi"], version = "^0.77.0"}
fastapi = "^0.68.1"
sqlmodel = "^0.0.4"


[tool.poetry.dev-dependencies]
black = "^21.8b0"
flake8 = "^3.9.2"
flake8-black = "^0.2.3"
flake8-bugbear = "^21.4.3"
mypy = "^0.910"
pre-commit = "^2.15.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
